Restructured the package structure of the app
Logically grouping related files makes it easy for developer to follow
what is going on in the project therefore making it easy to handle and
issues that may arise.

The structure of the project is logically set as follows

root package:
    - base: this package should contain code relating to the base feature
    of the app
    - di: this package should contain code relating to Dependency Injection
    - model: this package should contain models use in the app, this
    package can also be moved to feature specific packages
    - news: this is a feature specific package which contains feature
    specific code. There is a sub-packaged named "list" which contains
    views (UI related code) and viewmodel packages. A new sub-package can be
    add to this feature package i.e. "details" package which would contain
    news details code, with its view and viewmodel packages.

Feature packages can be added to the root package.
