+++ app/src/main/AndroidManifest.xml
@@ -8,11 +8,11 @@
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:name=".NewsApplication"
+        android:name=".di.NewsApplication"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.news.NewsActivity">
+        <activity android:name="com.android.nikhil.worldnow.news.list.view.NewsActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
