+++ app/src/main/java/com/android/nikhil/worldnow/di/component/AppComponent.kt
@@ -1,9 +1,9 @@
 package com.android.nikhil.worldnow.di.component
 
 import android.content.Context
-import com.android.nikhil.worldnow.NewsApplication
+import com.android.nikhil.worldnow.di.NewsApplication
 import com.android.nikhil.worldnow.di.module.ActivityBindingModule
-import com.android.nikhil.worldnow.di.module.NetworkModule
+import com.android.nikhil.worldnow.di.module.ServiceModule
 import com.android.nikhil.worldnow.di.module.ViewModelFactoryModule
 import dagger.BindsInstance
 import dagger.Component
@@ -14,7 +14,7 @@ import javax.inject.Singleton
 @Singleton
 @Component(modules = [
   AndroidSupportInjectionModule::class,
-  NetworkModule::class,
+  ServiceModule::class,
   ViewModelFactoryModule::class,
   ActivityBindingModule::class
 ])
