+++ app/src/main/java/com/android/nikhil/worldnow/di/module/ActivityBindingModule.kt
@@ -1,6 +1,6 @@
 package com.android.nikhil.worldnow.di.module
 
-import com.android.nikhil.worldnow.ui.news.NewsActivity
+import com.android.nikhil.worldnow.news.list.view.NewsActivity
 import dagger.Module
 import dagger.android.ContributesAndroidInjector
