+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/app_bar_scrolling_view_behavior"
-    tools:context=".ui.news.NewsActivity"
+    tools:context=".news.list.view.NewsActivity"
     tools:showIn="@layout/activity_main">
 
     <android.support.v7.widget.RecyclerView
